deploy/requirements.txt
numpy
ppqi
pillow
fastapi[all]